+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and `Removed`.
 - Fix a bug where users upgrading from an older version of Ajour might have incorrect
   behavior when trying to resize a column in the Catalog
 - Ignored addons are now sorted correctly again
+- Fingerprint and addon cache entries are now properly deleted when the addon folder
+  is missing from the filesystem
 
 ## [0.5.4] - 2020-12-07